Refactor quality control app layout
